Splitting the GetSemanticVersion.msbuild script into CalculateSemanticVersion.msbuild and GetSemanticVersion.msbuild.
The first script calls GitHubVersion to determine the semantic version and writes that to a file. Then the second script can be used to parse that file to get the version numbers where they are needed.
This ensures that we only get the version once during the build process which prevents while tagging.

references #30